agregado enemigo fijo
-Se agrega "Enemy_static" a los prefabs.
-Desde gameManager se generan 2 tipos de enemy (fijo y no fijo) al azar.
-Según dificultad aumenta la frecuencia en que se generan.
-Enemy y Enemy_static tienen "damage" diferente. El valor tomarlo desde EnemyController.
