+++ Assets/Prefabs/Characters/Enemy_static.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4fe042c65aa5334469d4a2543f03deb1
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
