+++ Assets/Scripts/Control/Enemy.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 42d126de68752fd48ac0ab5843e9f281
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
