+++ ProjectSettings/TagManager.asset
@@ -7,7 +7,6 @@ TagManager:
   - Enemy
   - Goal
   - Pickup
-  - FixedEnemy
   - LeftBound
   - RightBound
   - UpperBound
